+++ .gitignore
@@ -32,3 +32,5 @@ scripts
 
 # sqlite3
 *.db
+
+thoughts/
