+++ internal/repositories/postgresql/user.go
@@ -1,21 +1,15 @@
-package repositories
+package postgresql
 
 import (
 	"context"
 	"database/sql"
 	"errors"
-	"time"
 
 	model "github.com/KalessinD/gophermart/internal/models"
-	"github.com/KalessinD/gophermart/internal/services/db/pgerrors"
 	"github.com/jackc/pgx/v5/pgconn"
 )
 
 const (
-	RetryingAttempts  = 3
-	RetryingDelay     = 100 * time.Millisecond
-	RetryingDelayStep = 200 * time.Millisecond
-
 	PsqlUsersSchema = "gophermart"
 	PsqlUserTable   = "users"
 
@@ -28,9 +22,6 @@ const (
 )
 
 type (
-	txWrapperFunc func(tx *sql.Tx) error
-	wrapperFunc   func(ctx context.Context) (*sql.Row, error)
-
 	SQLStorage struct {
 		db *sql.DB
 	}
@@ -51,7 +42,7 @@ func (r *SQLStorage) Ping(ctx context.Context) error {
 }
 
 func (r *SQLStorage) GetUser(ctx context.Context, login string) (*model.User, error) {
-	row, err := r.withRetry(ctx, func(ctx context.Context) (*sql.Row, error) {
+	row, err := withRetry(ctx, func(ctx context.Context) (*sql.Row, error) {
 		row := r.db.QueryRowContext(ctx, QuerySelectUser, login)
 		if row.Err() != nil {
 			return nil, row.Err()
@@ -72,7 +63,7 @@ func (r *SQLStorage) GetUser(ctx context.Context, login string) (*model.User, er
 }
 
 func (r *SQLStorage) AddUser(ctx context.Context, user *model.User) error {
-	return r.withTxRetry(ctx, func(tx *sql.Tx) error {
+	return withTxRetry(ctx, r.db, func(tx *sql.Tx) error {
 		if err := r.addUserTx(ctx, tx, user); err != nil {
 			return err
 		}
@@ -92,54 +83,4 @@ func (r *SQLStorage) addUserTx(ctx context.Context, tx *sql.Tx, user *model.User
 		return err
 	}
 	return nil
-}
-
-func (r *SQLStorage) withRetry(ctx context.Context, action wrapperFunc) (*sql.Row, error) {
-	var err error
-
-	for attempts := 0; attempts < RetryingAttempts; attempts++ {
-		obj, err := action(ctx)
-
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			if pgerrors.ClassifyPgError(pgErr) == pgerrors.Retriable {
-				time.Sleep(time.Millisecond * 100)
-				continue
-			}
-		}
-
-		return obj, err
-	}
-
-	return nil, err
-}
-
-func (r *SQLStorage) withTxRetry(ctx context.Context, action txWrapperFunc) error {
-	var err error
-
-	for attempts := 0; attempts < RetryingAttempts; attempts++ {
-		tx, err := r.db.BeginTx(ctx, nil)
-		if err != nil {
-			return err
-		}
-
-		err = action(tx)
-		if err == nil {
-			return tx.Commit()
-		}
-
-		_ = tx.Rollback()
-
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			if pgerrors.ClassifyPgError(pgErr) == pgerrors.Retriable {
-				time.Sleep(time.Millisecond * 100)
-				continue
-			}
-		}
-
-		return err
-	}
-
-	return err
 }
